step2 피드백 반영 - QnaServiceTest 수정(실제 데이터로 테스트) - 트랜잭션 관련 수정 - 사용자 확인 메소드 수정

+++ src/main/java/nextstep/service/QnaService.java
@@ -2,6 +2,7 @@ package nextstep.service;
 
 import nextstep.UnAuthorizedException;
 import nextstep.domain.*;
+import nextstep.dto.QuestionDTO;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.data.domain.Pageable;
@@ -33,18 +34,26 @@ public class QnaService {
 
     public Question findById(long id) {
         return questionRepository.findById(id)
-                .orElseThrow(EntityNotFoundException::new);
+                .orElseThrow(() -> new EntityNotFoundException("해당 질문글을 찾을 수 없습니다."));
+    }
+
+    public Answer findAnswerById(long id) {
+        return answerRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("해당 답변을 찾을 수 없습니다."));
     }
 
     @Transactional
     public Question update(User loginUser, long id, Question updatedQuestion) {
         checkQuestionOwner(id, loginUser);
-        return questionRepository.save(updatedQuestion);
+        Question origin = findById(id);
+        origin.update(origin, updatedQuestion);
+        return origin;
     }
 
     @Transactional
     public void deleteQuestion(User loginUser, long questionId) {
-        Question question = checkQuestionOwner(questionId, loginUser);
+        checkQuestionOwner(questionId, loginUser);
+        Question question = findById(questionId);
         questionRepository.delete(question);
     }
 
@@ -56,30 +65,40 @@ public class QnaService {
         return questionRepository.findAll(pageable).getContent();
     }
 
+    @Transactional
     public Answer addAnswer(User loginUser, long questionId, String contents) {
-        Question question = questionRepository.findById(questionId)
-                .orElseThrow(EntityNotFoundException::new);
+        Question question = findById(questionId);
         Answer answer = new Answer(loginUser, contents);
         question.addAnswer(answer);
         return answerRepository.save(answer);
     }
 
+    @Transactional
     public Answer deleteAnswer(User loginUser, long id) {
-        Answer answer = answerRepository.findById(id)
-                .filter(a -> a.isOwner(loginUser))
-                .orElseThrow(() -> new EntityNotFoundException("해당 답변을 찾을 수 없습니다."));
+        checkAnswerOwner(id, loginUser);
+        Answer answer = findAnswerById(id);
         answerRepository.delete(answer);
         return answer;
     }
 
-    private Question checkQuestionOwner(long questionId, User loginUser) {
-        Question question = questionRepository.findById(questionId)
-                .orElseThrow(() -> new EntityNotFoundException("질문글을 찾을 수 없습니다."));
+    @Transactional
+    public QuestionDTO findQuestionAndAnswerById(long questionId) {
+        Question question = findById(questionId);
+        int answerSize = question.getAnswers().size();
+        return new QuestionDTO(question, answerSize);
+    }
 
+    private void checkQuestionOwner(long questionId, User loginUser) {
+        Question question = findById(questionId);
         if(!question.isOwner(loginUser)) {
             throw new UnAuthorizedException("해당 사용자가 작성한 글이 아닙니다.");
         }
+    }
 
-        return question;
+    private void checkAnswerOwner(long id, User loginUser) {
+        Answer answer = findAnswerById(id);
+        if(!answer.isOwner(loginUser)) {
+            throw new UnAuthorizedException("해당 사용자가 작성한 답변이 아닙니다.");
+        }
     }
 }
